+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 2022.xx.xx
-* Minor bug fix
+* Speed-up of diagonalization process with ELPA for spin-symmetric mode.
+* Minor bug fixes.
 
 # 2022.04.05
 * Update of SLDAE functional (redefinition of the effective mass, now A=alpha; added a new flag that forces the effective mass to be 1.0).