+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 2022.xx.xx
+* Speed-up of diagonalization process with ELPA for spin-symmetric mode.
+* Minor bug fixes.
+
 # 2022.04.05
 * Update of SLDAE functional (redefinition of the effective mass, now A=alpha; added a new flag that forces the effective mass to be 1.0).
 * Redefinition of delta_ext, now it is consistent with https://arxiv.org/abs/2201.07626.